test: Move create_test_tx_sender and create_bumpable_tx to common + write initial TxSenderReorgBehavior test.
