+++ core/src/tx_sender.rs
@@ -766,7 +766,6 @@ impl TxSender {
         new_fee_rate: FeeRate,
         current_tip_height: u32,
     ) -> Result<()> {
-        tracing::info!("Trying to send unconfirmed txs with new fee rate: {new_fee_rate:?}, current tip height: {current_tip_height:?}");
         let txs = self
             .db
             .get_sendable_txs(None, new_fee_rate, current_tip_height)
@@ -1096,9 +1095,6 @@ mod tests {
     use crate::builder::script::{CheckSig, SpendableScript};
     use crate::builder::transaction::TransactionType;
     use crate::{database::Database, test::common::*};
-    use bitcoin::secp256k1::SecretKey;
-    use bitcoin::transaction::Version;
-    use secp256k1::rand;
     use std::result::Result;
     use std::sync::Arc;
 
@@ -1108,94 +1104,6 @@ mod tests {
         ) -> Result<sqlx::Transaction<'_, sqlx::Postgres>, BridgeError> {
             self.db.begin_transaction().await
         }
-    }
-
-    async fn create_test_tx_sender(
-        rpc: ExtendedRpc,
-    ) -> (TxSender, ExtendedRpc, Database, Actor, bitcoin::Network) {
-        let sk = SecretKey::new(&mut rand::thread_rng());
-        let network = bitcoin::Network::Regtest;
-        let actor = Actor::new(sk, None, network);
-
-        let config = create_test_config_with_thread_name().await;
-
-        let db = Database::new(&config).await.unwrap();
-
-        let tx_sender = TxSender::new(
-            actor.clone(),
-            rpc.clone(),
-            db.clone(),
-            "tx_sender".into(),
-            network,
-        );
-
-        (tx_sender, rpc, db, actor, network)
-    }
-
-    async fn create_bumpable_tx(
-        rpc: &ExtendedRpc,
-        signer: &Actor,
-        network: bitcoin::Network,
-        fee_paying_type: FeePayingType,
-    ) -> Result<Transaction, BridgeError> {
-        let (address, spend_info) =
-            builder::address::create_taproot_address(&[], Some(signer.xonly_public_key), network);
-
-        let amount = Amount::from_sat(100000);
-        let outpoint = rpc.send_to_address(&address, amount).await?;
-        rpc.mine_blocks(1).await?;
-
-        let version = match fee_paying_type {
-            FeePayingType::CPFP => Version::non_standard(3),
-            FeePayingType::RBF => Version::TWO,
-        };
-
-        let mut builder = TxHandlerBuilder::new(TransactionType::Dummy)
-            .with_version(version)
-            .add_input(
-                NormalSignatureKind::OperatorSighashDefault,
-                SpendableTxIn::new(
-                    outpoint,
-                    TxOut {
-                        value: amount,
-                        script_pubkey: address.script_pubkey(),
-                    },
-                    vec![],
-                    Some(spend_info),
-                ),
-                SpendPath::KeySpend,
-                DEFAULT_SEQUENCE,
-            )
-            .add_output(UnspentTxOut::from_partial(TxOut {
-                value: amount - builder::transaction::anchor_output().value,
-                script_pubkey: address.script_pubkey(), // TODO: This should be the wallet address, not the signer address
-            }))
-            .add_output(UnspentTxOut::from_partial(
-                builder::transaction::anchor_output(),
-            ))
-            .finalize();
-
-        let sighash_type = match fee_paying_type {
-            FeePayingType::CPFP => bitcoin::TapSighashType::Default,
-            FeePayingType::RBF => bitcoin::TapSighashType::AllPlusAnyoneCanPay,
-        };
-
-        let sighash = builder.calculate_pubkey_spend_sighash(0, sighash_type)?;
-        let signature = signer.sign_with_tweak_data(
-            sighash,
-            builder::sighash::TapTweakData::KeyPath(None),
-            None,
-        )?;
-        builder.set_p2tr_key_spend_witness(
-            &bitcoin::taproot::Signature {
-                signature,
-                sighash_type,
-            },
-            0,
-        )?;
-
-        let tx = builder.get_cached_tx().clone();
-        Ok(tx)
     }
 
     #[tokio::test]
